Renamed eval to simulator
