+++ .gitignore
@@ -27,6 +27,7 @@ share/python-wheels/
 *.egg
 MANIFEST
 .vscode/
+release.sh
 .DS_Store
 
 # PyInstaller
